Config and QoL updates
Updated the order of existing configuration and what gets displayed for them. Added a configuration for HTTP timeout, automatically refreshing, and the period for automatic refreshing. The Google Sheet HTTP requests will now only occur after you click on the Clan Events panel, before it was doing this every time a configuration was changed and on startup, which made the Runelite startup loading screen take a lot longer.