Update default http timeout
Update the default http timeout value to 1s seeing as even 1 second is typically enough and higher values make it more likely to cause problems due to the game hanging during the request. Also, change the max value to 5 seconds because it should never take more than 5 seconds for an http request.